requirements.txt
Flask>=3.0
Flask-Login
Flask-SQLAlchemy
PyMySQL
python-dotenv
pyotp
qrcode
Pillow
